requirements.txt
numpy==1.26.4
opencv-python==4.8.1.78
face_recognition==1.3.0
pymongo
python-dotenv
pandas
openpyxl
